Allow plugins opportunity to fail loading
A plugin can now abort from the Initialize function, this
will allow the plugin to fail loading and be unloaded without
having to do the DoCommand("/plugin foo unload") method.

This also means that plugin errors will be a little bit more
discriptive as failure to load a plugin can now specify an
error message.