Add basic dependency support
Declare plugin dependencies by implementing the ConfigurePlugin
function and filling the PLUGINCONFIG structure with the relevant
information.

Loading a plugin with unmet dependencies will fail unless the
"force" option is provided:

   /plugin mq2foo force

Unloading a plugin that is being depended on by other plugins will
also fail unless the "force" option is provided:

  /plugin mq2foo unload force

Added new macro CreatePluginClass which can be used to declare the
plugin class for a new style plugin, so that MQ2 can initialize it.