Create MQ2PluginInfo class, revert MQPLUGIN struct
Changed plugin structuring so that MQPLUGIN struct exists how it
originally was (minus the function pointers). The plugin handling
code will maintain both the new list of MQ2PluginInfo objects, and
the old list of MQPLUGIN objects. This is simply for backwards
compatibility with plugins that are written to access the plugin
list directly.

Update the Plugin TLO and Plugin datatype to use the new types.