Detect visual studio toolchain mismatch in mq2 plugins
Added a change to detect the visual studio version used to compile
plugins, to ensure that all the plugins are being built with the
same toolchain. This is to help avoid unexpected stability problems
due to mismatched compilers and library versions.

In the future I will add an option that allows a plugin to opt out
of this check.