Add new feature to allow extending existing data types
New functions:
- AddMQ2TypeExtension(const char* szName, MQ2Type* extension)
- RemoveMQ2TypeExtension(const char* szName, MQ2Type* extension)

These functions allow you to extend an existing type by providing the
name of the type you want to extend and the custom MQ2Type instance that
does the extending.

Currently this works only for members. It will check for members of the
extension type before checking members of the base type. Its sort of
like inheritance, but doesn't require modifying the original type.

See MQ2Test/MQ2Test.cpp for an example.

Also made some improvements and optimizations to the MQ2DataAPI code.